Improvements on new command
